add split strategy
